+++ clients/get_service_instance_by_name.go
@@ -35,7 +35,12 @@ func GetServiceInstanceByName(cliConnection plugin.CliConnection, spaceGUID stri
 		}
 
 		for _, serviceInstance := range responseObject.Resources {
-			serviceInstances = append(serviceInstances, models.CFServiceInstance{Name: *serviceInstance.Entity.Name, GUID: serviceInstance.Metadata.GUID, UpdatedAt: serviceInstance.Metadata.UpdatedAt})
+			serviceInstances = append(serviceInstances, models.CFServiceInstance{
+				Name:          *serviceInstance.Entity.Name,
+				GUID:          serviceInstance.Metadata.GUID,
+				UpdatedAt:     serviceInstance.Metadata.UpdatedAt,
+				LastOperation: *serviceInstance.Entity.LastOperation,
+			})
 		}
 		nextURL = responseObject.NextURL
 	}