+++ clients/models/cf_service_instance.go
@@ -2,7 +2,8 @@ package models
 
 // CFServiceInstance Cloud Foundry service instance
 type CFServiceInstance struct {
-	Name      string
-	GUID      string
-	UpdatedAt string
+	Name          string
+	GUID          string
+	UpdatedAt     string
+	LastOperation CFLastOperation
 }