+++ commands/html5_push_command.go
@@ -478,7 +478,7 @@ func (c *PushCommand) PushHTML5Applications(appPaths []string, appHostGUID strin
 
 	// Delete temporarry service keys
 	log.Tracef("Deleting temporarry service key: '%s'\n", serviceKey.Name)
-	err = clients.DeleteServiceKey(c.CliConnection, serviceKey.GUID)
+	err = clients.DeleteServiceKey(c.CliConnection, serviceKey.GUID, maxRetryCount)
 	if err != nil {
 		ui.Failed("Could not delete service key '%s' : %+v", serviceKey.Name, err)
 		return Failure